Add files via upload
I tried to reconstruct a model based on the txt file points3D. First i did it with only points in coordinate, then i tried to use the track and error which are given in the file to do it into lines but it didnt workout. Can somebody maybe take a look at this for the 3D reconstruct? It may bring u some inspiritions. The point 3D reconstruct is in a txt file also.